+++ template/backend/src/core/index.ts
