+++ template/backend/src/database/schema/index.ts
@@ -1,1 +1,4 @@
-export * from './auth'
+export * from './accounts'
+export * from './sessions'
+export * from './users'
+export * from './verifications'
